requirement.txt
fastapi
sqlalchemy
uvicorn
bcrypt
passlib
python-jose